+++ .circleci/config.yml
@@ -49,7 +49,7 @@ jobs:
           name: run tests
           command: |
            . venv/bin/activate
-           pwd && ls -al && chmod 755 ./l_geckodriver  && chmod 755 ./l_chromedriver && PATH=$PATH:/home/circleci/repo && pytest 
+           chmod 755 ./l_geckodriver  && chmod 755 ./l_chromedriver && PATH=$PATH:/home/circleci/repo && pytest -n 4
 
 
       - store_artifacts:
