+++ fibo/conftest.py
@@ -2,11 +2,13 @@ import pytest
 import requests
 from dev_classes.phase4_UI.jira_pages import LoginPO
 from dev_classes.phase4_UI.projectWd import ProjectWD
-from dev_classes.phase4_UI.properties4 import user_name, user_pass, pytestScope
 from dev_classes.phase2_API.json_fixtures import *
 from dev_classes.phase2_API.properties_for_phase2 import *
+from dev_classes.phase4_UI.properties4 import user_name, user_pass, pytestScope
 
 
+# if (pytestScope='function')  -> Webdriver will init/teardown per... suddenly function!
+# if (pytestScope='class')  -> Webdriver will init/teardown per class
 @pytest.fixture(scope=pytestScope)
 def wd():
     dr = ProjectWD()
